requirements.txt
py-cord==2.4.1
python-dotenv==0.21.0
peewee==3.16.2
peewee-migrations==0.3.32
pymysql==1.0.2
cryptography==38.0.4
pynacl==1.5.0